requirements.txt
# Interactive inference UI
pygame>=2.5.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
av==15.1.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmake==4.1.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.1.1
deepdiff==8.6.1
diffusers==0.35.2
dill==0.4.0
draccus==0.10.0
einops==0.8.1
Farama-Notifications==0.0.4
feetech-servo-sdk==1.0.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gymnasium==1.2.2
hf-xet==1.2.0
hf_transfer==0.1.9
huggingface-hub==0.35.3
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inquirerpy==0.3.4
Jinja2==3.1.6
jsonlines==4.0.0
kiwisolver==1.4.9
lerobot==0.4.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mergedeep==1.3.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
pfzy==0.3.4
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pynput==1.8.1
pyobjc-core==12.1; sys_platform == "darwin"
pyobjc-framework-ApplicationServices==12.1; sys_platform == "darwin"
pyobjc-framework-Cocoa==12.1; sys_platform == "darwin"
pyobjc-framework-CoreText==12.1; sys_platform == "darwin"
pyobjc-framework-Quartz==12.1; sys_platform == "darwin"
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyyaml-include==1.4.1
regex==2025.11.3
requests==2.32.5
rerun-sdk==0.26.2
safetensors==0.7.0
sentry-sdk==2.47.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
termcolor==3.2.0
toml==0.10.2
torch==2.7.1
torchcodec==0.5
torchvision==0.22.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
wandb==0.21.4
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
pygame==2.6.1

# Reward bead suite UI
streamlit==1.29
streamlit-drawable-canvas>=0.9